Feat : Chased down warnings and converted float to double for accuracy
Also added function to remove all contact with two rigidbodies

+++ TEST2/PPlane.h
@@ -4,16 +4,16 @@
 
 class PPlane : public Primitive {
 public:
-	PPlane(Vector3D normal, RigidBody* RB, Matrix4 offset, float HalfX, float HalfY) : Primitive(RB, offset), normal(normal), halfsizeX(HalfX), halfsizeY(HalfY){
+	PPlane(Vector3D normal, RigidBody* RB, Matrix4 offset, double HalfX, double HalfY) : Primitive(RB, offset), normal(normal), halfsizeX(HalfX), halfsizeY(HalfY){
 	}
 
 	// vecteur perpendiculaire à la surface du plan
 	Vector3D normal;
 
 	//Halfsize
-	float halfsizeX;
-	float halfsizeY;
+	double halfsizeX;
+	double halfsizeY;
 
 	// décalage du plan par rapport à l'origine
-	float offset; // Déjà dans Primitive ?
+	double offset; // Déjà dans Primitive ?
 };
